refactor: Reduce config build requirements
