+++ osquery/core/plugins/BUCK
@@ -10,7 +10,7 @@ load("//tools/build_defs/oss/osquery:native.bzl", "osquery_target")
 osquery_cxx_library(
     name = "plugins",
     srcs = glob(["*.cpp"]),
-    header_namespace = "osquery/plugins",
+    header_namespace = "osquery/core/plugins",
     exported_headers = glob(["*.h"]),
     visibility = ["PUBLIC"],
     deps = [
