+++ osquery/core/plugins/sql_plugin.h
@@ -18,9 +18,9 @@
 
 #pragma once
 
+#include <osquery/core/plugins/plugin.h>
 #include <osquery/core/sql/column.h>
 #include <osquery/core/sql/query_data.h>
-#include <osquery/plugins/plugin.h>
 #include <osquery/utils/status/status.h>
 
 namespace osquery {
