+++ osquery/dispatcher/CMakeLists.txt
@@ -11,13 +11,13 @@ function(osqueryDispatcherMain)
   endif()
 
   generateOsqueryDispatcher()
+  generateOsqueryDispatcherScheduler()
 endfunction()
 
 function(generateOsqueryDispatcher)
   add_osquery_library(osquery_dispatcher EXCLUDE_FROM_ALL
     dispatcher.cpp
     distributed_runner.cpp
-    scheduler.cpp
   )
 
   target_link_libraries(osquery_dispatcher PUBLIC
@@ -26,21 +26,39 @@ function(generateOsqueryDispatcher)
     osquery_logger_datalogger
     osquery_process
     osquery_profiler
-    osquery_sql
     osquery_utils
     osquery_utils_conversions
-    plugins_config_parsers
     thirdparty_googletest_headers
   )
 
   set(public_header_files
     distributed_runner.h
-    scheduler.h
   )
 
   generateIncludeNamespace(osquery_dispatcher "osquery/dispatcher" "FILE_ONLY" ${public_header_files})
 
   add_test(NAME osquery_dispatcher_tests-test COMMAND osquery_dispatcher_tests-test)
+endfunction()
+
+function(generateOsqueryDispatcherScheduler)
+  add_osquery_library(osquery_dispatcher_scheduler EXCLUDE_FROM_ALL
+    scheduler.cpp
+  )
+
+  target_link_libraries(osquery_dispatcher_scheduler PUBLIC
+    osquery_cxx_settings
+    osquery_core
+    osquery_dispatcher
+    plugins_config_parsers
+    thirdparty_googletest_headers
+  )
+
+  set(public_header_files
+    scheduler.h
+  )
+
+  generateIncludeNamespace(osquery_dispatcher_scheduler "osquery/dispatcher" "FILE_ONLY" ${public_header_files})
+
   add_test(NAME osquery_dispatcher_tests_scheduler-test COMMAND osquery_dispatcher_tests_scheduler-test)
 endfunction()
