+++ osquery/events/linux/udev.h
@@ -87,26 +87,6 @@ class UdevEventPublisher
 
   Status run() override;
 
-  /**
-   * @brief Return a string representation of a udev property.
-   *
-   * @param device the udev device pointer.
-   * @param property the udev property identifier string.
-   * @return string representation of the property or empty if null.
-   */
-  static std::string getValue(struct udev_device* device,
-                              const std::string& property);
-
-  /**
-   * @brief Return a string representation of a udev system attribute.
-   *
-   * @param device the udev device pointer.
-   * @param property the udev system attribute identifier string.
-   * @return string representation of the attribute or empty if null.
-   */
-  static std::string getAttr(struct udev_device* device,
-                             const std::string& attr);
-
  private:
   /// udev handle (socket descriptor contained within).
   struct udev* handle_{nullptr};
@@ -125,4 +105,4 @@ class UdevEventPublisher
   /// Helper function to create an EventContext using a udev_device pointer.
   UdevEventContextRef createEventContextFrom(struct udev_device* device);
 };
-}
+} // namespace osquery
