+++ osquery/extensions/extensions.cpp
@@ -524,52 +524,6 @@ Status startExtension(const std::string& manager_path,
           << sdk_version << ") registered";
   return Status(0, std::to_string(uuid));
 }
-Status ExternalSQLPlugin::query(const std::string& query,
-                                QueryData& results,
-                                bool use_cache) const {
-  static_cast<void>(use_cache);
-  // Make sure the extension path exists, and is writable.
-  auto status = extensionPathActive(FLAGS_extensions_socket);
-  if (!status.ok()) {
-    return status;
-  }
-
-  try {
-    ExtensionManagerClient client(FLAGS_extensions_socket);
-    status = client.query(query, results);
-  } catch (const std::exception& e) {
-    return Status(1, "Extension call failed: " + std::string(e.what()));
-  }
-
-  return status;
-}
-
-Status ExternalSQLPlugin::getQueryColumns(const std::string& query,
-                                          TableColumns& columns) const {
-  // Make sure the extension path exists, and is writable.
-  auto status = extensionPathActive(FLAGS_extensions_socket);
-  if (!status.ok()) {
-    return status;
-  }
-
-  QueryData qd;
-  try {
-    ExtensionManagerClient client(FLAGS_extensions_socket);
-    status = client.getQueryColumns(query, qd);
-  } catch (const std::exception& e) {
-    return Status(1, "Extension call failed: " + std::string(e.what()));
-  }
-
-  // Translate response map: {string: string} to a vector: pair(name, type).
-  for (const auto& column : qd) {
-    for (const auto& col : column) {
-      columns.push_back(std::make_tuple(
-          col.first, columnTypeName(col.second), ColumnOptions::DEFAULT));
-    }
-  }
-
-  return status;
-}
 
 Status pingExtension(const std::string& path) {
   if (FLAGS_disable_extensions) {
