+++ osquery/sql/BUCK
@@ -14,6 +14,7 @@ osquery_cxx_library(
     srcs = [
         "dynamic_table_row.cpp",
         "sql.cpp",
+        "sql_sqlite_plugin.cpp",
         "sqlite_encoding.cpp",
         "sqlite_filesystem.cpp",
         "sqlite_hashing.cpp",
