+++ osquery/sql/sqlite_carver.cpp
