+++ plugins/config/update.cpp
@@ -6,7 +6,7 @@
  *  the LICENSE file found in the root directory of this source tree.
  */
 
-#include <osquery/config/config.h>
+#include <osquery/config/config_plugin.h>
 #include <osquery/registry_factory.h>
 
 namespace osquery {
@@ -35,4 +35,4 @@ class UpdateConfigPlugin : public ConfigPlugin {
 };
 
 REGISTER(UpdateConfigPlugin, "config", "update");
-}
+} // namespace osquery
